requirements.txt
kaggle>=1.7.4.5
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0.1
pytest>=7.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.11.0
